Update import paths after recent move.
