+++ src/ogd/common/storage/interfaces/BigQueryInterface.py
@@ -18,7 +18,7 @@ from ogd.common.models.enums.IDMode import IDMode
 from ogd.common.models.enums.FilterMode import FilterMode
 from ogd.common.models.enums.VersionType import VersionType
 from ogd.common.storage.interfaces.Interface import Interface
-from ogd.common.storage.connectors.BigQueryConnector import BigQueryConnector
+from ogd.common.storage.connectors.cores.BigQueryConnector import BigQueryConnector
 from ogd.common.utils.Logger import Logger
 
 AQUALAB_MIN_VERSION : Final[float] = 6.2
