+++ src/ogd/common/storage/interfaces/Interface.py
@@ -19,7 +19,7 @@ from ogd.common.models.SemanticVersion import SemanticVersion
 from ogd.common.configs.GameStoreConfig import GameStoreConfig
 from ogd.common.schemas.tables.EventTableSchema import EventTableSchema
 from ogd.common.schemas.tables.FeatureTableSchema import FeatureTableSchema
-from ogd.common.storage.connectors.StorageConnector import StorageConnector
+from ogd.common.storage.connectors.cores.StorageConnector import StorageConnector
 from ogd.common.utils.Logger import Logger
 
 class Interface(abc.ABC):
