+++ tests/cases/storage/interfaces/test_BigQueryInterface.py
@@ -14,7 +14,7 @@ from ogd.common.configs.TestConfig import TestConfig
 from ogd.common.models.enums.IDMode import IDMode
 from ogd.common.models.enums.FilterMode import FilterMode
 from ogd.common.schemas.tables.EventTableSchema import EventTableSchema
-from ogd.common.storage.connectors.BigQueryConnector import BigQueryConnector
+from ogd.common.storage.connectors.cores.BigQueryConnector import BigQueryConnector
 from ogd.common.storage.interfaces.BigQueryInterface import BigQueryInterface, ParamaterizedClause
 from ogd.common.utils.Logger import Logger
 # import locals
